hopefully right

+++ synonyms.py
@@ -124,7 +124,7 @@ def run_similarity_test(filename, semantic_descriptors, similarity_fn):
         if guess == ans:
             score += 1
     
-    return((score/(len(list1)))*100)
+    return float((score/(len(list1)))*100)
 
 
 #if __name__ == "__main__":
